fix: BE-017 update imports and exports
