+++ src/client.ts
@@ -2,6 +2,4 @@ import {
   PrismaClient,
 } from '@prisma/client';
 
-const prisma = new PrismaClient();
-
-export { prisma };
+export const prisma = new PrismaClient();
