+++ src/graphql/models/Query.ts
@@ -4,7 +4,7 @@
  *  login(email: String, password: String) AuthMessage!
  *  logout: Void
  *  findOneTask(id: String!) Task!
- */ 
+ */
 export const Query = `
   type Query {
     hello: Message!
