--- src/graphql/models/User.ts
@@ -1,8 +1,0 @@
-export const User =`
-type User {
-  id: ID!
-  name: String
-  email: String
-  password: String
-  activated: Boolean
-}`;
