+++ src/graphql/models/hello/Types.ts
@@ -1,0 +1,10 @@
+/**
+ * Message type definition
+ */
+export const Message = `
+ type Message {
+   message: String!
+ }
+ `;
+
+export const RandomTypes: string[] = [Message];
