+++ src/graphql/models/index.ts
@@ -1,5 +1,12 @@
-import { User } from './User';
+import { AuthTypes } from './auth/Types';
+import { RandomTypes } from './hello/Types';
+import { Mutation } from './Mutations';
+import { Query } from './Query';
 
-export default [
-  User
+export const Types = [...AuthTypes, ...RandomTypes];
+
+export const typeDefs = [
+    Mutation,
+    Query,
+    ...Types,
 ];
