--- src/graphql/utils/AuthPayload.ts
@@ -1,5 +1,0 @@
-export const AuthPayload = `
-type AuthPayload {
-  user: User!
-}
-`;
