--- src/graphql/utils/Message.ts
@@ -1,5 +1,0 @@
-export const Message = `
-type Message {
-  message: String!
-}
-`;
