--- src/graphql/utils/index.ts
@@ -1,11 +1,0 @@
-import { AuthPayload } from './AuthPayload';
-import { Message } from './Message';
-import { Mutation } from './Mutation';
-import { Query } from './Query';
-
-export default [
-  Mutation,
-  Query,
-  AuthPayload,
-  Message
-];
