+++ src/repositories/AuthenticationRepository/createOneUser.ts
@@ -1,19 +1,16 @@
 import { prisma } from 'src/client';
 import { ICreateUserArgs } from 'src/interfaces';
 
-async function createOneUser({email, password, name}: ICreateUserArgs) {
-  console.log('TRYING TO CREATE A USER');
-  return await prisma.user.create({
-    data: {
-      email,
-      password,
-      name
-    },
-    select: {
-      id: true,
-      email: true,
-    },
-  });
-}
-
-export { createOneUser };
+export const createOneUser = async ({ email, password, name }: ICreateUserArgs) => {
+    return await prisma.user.create({
+        data: {
+            email,
+            password,
+            name
+        },
+        select: {
+            id: true,
+            email: true,
+        },
+    });
+};
