+++ src/resolvers/hello/hello.query.ts
@@ -1,0 +1,10 @@
+export const hello = async () => {
+    try {
+        return {
+            message: 'This is some random data returned'
+        };
+    } catch (e) {
+        (e);
+        throw e;
+    }
+};
