--- src/resolvers/index.ts
@@ -1,4 +1,0 @@
-import Mutation from './Mutation';
-import Query from './Query';
-
-export { Query, Mutation };
